feat: Complete Phase 3 services consolidation (Flow + Message → Domain)
✅ Phase 3 Complete - Flow & Message Services to Domain

Migrated 9 additional services to Domain-Driven Design architecture,
bringing total domain completion to 95%.

## Services Migrated

### Flow Services (2)
1. FlowDataIntegrityChecker → domain/flows/integrity/
2. FlowEventBroadcaster → domain/flows/events/

### Message Services (7 → 4 subdomains)
1. MessageService → domain/messaging/core/
2. MessageBaseService → domain/messaging/core/
3. MessageFactory → domain/messaging/core/
4. MessageScheduler → domain/messaging/scheduling/
5. MessageSender → domain/messaging/delivery/
6. IdempotentMessageSender → domain/messaging/delivery/
7. WhatsAppService → domain/messaging/whatsapp/

## New Domain Structure

app/domain/
├── flows/
│   ├── integrity/ (NEW - data corruption detection)
│   └── events/ (NEW - WebSocket broadcasting)
└── messaging/ (NEW - complete domain, 4 subdomains)
    ├── core/
    ├── scheduling/
    ├── delivery/
    └── whatsapp/

## Backward Compatibility

✅ All old import paths work via deprecation adapters
✅ DeprecationWarnings guide developers
✅ Zero test failures
✅ Zero breaking changes

## Architecture Benefits

- Clear domain boundaries and logical organization
- Single Responsibility Principle enforced
- Reduced code duplication
- Better discoverability and maintainability
- Full DDD compliance

## Impact

- Services migrated: 9 (Phase 3) + 8 (Phase 2) + 12 (Phase 1) = 29 total
- Domain completion: 78% → 95% (+17 points)
- Files organized: 44 domain files created
- Adapters created: 9 backward-compatible adapters
- Breaking changes: 0

## Documentation

- Full migration guide: docs/migrations/PHASE_3_SERVICES_CONSOLIDATION.md
- Import examples: Old vs New patterns
- Architecture diagrams included

## Validation

✅ All Python files syntax validated
✅ All imports resolved
✅ Zero circular dependencies
✅ Tests passing
