+++ backend-hormonia/app/domain/messaging/whatsapp/__init__.py
@@ -1,0 +1,4 @@
+"""WhatsApp integration services."""
+from .whatsapp_service import WhatsAppService
+
+__all__ = ["WhatsAppService"]
